Allow retrieving archived data dumps by providing a date
Credit to sans for the idea of returning an archive if a date is provided.

Also changed dumpManager to not maintain a dumpNodes map, instead there are no defaults for retrieve_iterator, and the high level methods pass specific tags. Makes it easy to collect similar dumps, such as archives and the multiple card dumps.